fix some bugs

+++ src/hikogui/GUI/gui_window_win32.hpp
@@ -47,7 +47,7 @@ private:
     gui_event mouse_button_event;
     utc_nanoseconds multi_click_time_point;
     point2 multi_click_position;
-    int multi_click_count;
+    uint8_t multi_click_count;
 
     void setOSWindowRectangleFromRECT(RECT aarectangle) noexcept;
